+++ attributes/default.rb
@@ -1,3 +1,1 @@
 default['dvo_user']['sumologic']['storage_class'] = 'standard'
-default['dvo_user']['sumologic']['accessID'] = 'surjan05DU9ut2'
-default['dvo_user']['sumologic']['accessKey'] = 'LlgS3jojH7LXJxv1UnbaLj1ZPXN38oWZXLDXJqRJ59rTA6xjVwQ0XrP08pf2o4YV'